Split up the LLVM setup code for crux-dbg
This also exports the overrides so that they could be reused elsewhere
